+++ adapter/spark/src/main/java/com/asml/apa/wta/spark/WtaPlugin.java
@@ -30,6 +30,7 @@ public class WtaPlugin implements SparkPlugin {
    * Returns the executor-side component of the plugin.
    *
    * @return The executor-side component initialised at startup
+   *
    */
   @Override
   public ExecutorPlugin executorPlugin() {