+++ adapter/spark/src/main/java/com/asml/apa/wta/spark/driver/WtaDriverPlugin.java
@@ -1,5 +1,7 @@
 package com.asml.apa.wta.spark.driver;
 
+import com.asml.apa.wta.core.model.Workload;
+import com.asml.apa.wta.spark.datasource.SparkDataSource;
 import java.util.HashMap;
 import java.util.Map;
 import org.apache.spark.SparkContext;
@@ -14,6 +16,12 @@ import org.apache.spark.api.plugin.PluginContext;
  */
 public class WtaDriverPlugin implements DriverPlugin {
 
+  private SparkContext sparkContext;
+
+  private SparkDataSource sparkDataSource;
+
+  private Workload.WorkloadBuilder workloadBuilder;
+
   /**
    * This method is called early in the initialization of the Spark driver.
    * Explicitly, it is called before the Spark driver's task scheduler is initialized. It is blocking.
@@ -23,15 +31,25 @@ public class WtaDriverPlugin implements DriverPlugin {
    * @param sCtx The current SparkContext.
    * @param pCtx Additional plugin-specific about the Spark application where the plugin is running.
    * @return Extra information provided to the executor
+   * @author Henry Page
+   * @since 1.0.0
    */
   @Override
   public Map<String, String> init(SparkContext sCtx, PluginContext pCtx) {
+    this.sparkContext = sCtx;
+    this.sparkDataSource = new SparkDataSource(sparkContext);
+    this.workloadBuilder = Workload.builder().startDate(sparkContext.startTime());
     return new HashMap<>();
   }
 
   /**
    * Gets called just before shutdown. Recommended that no spark functions are used here.
+   *
+   * @author Henry Page
+   * @since 1.0.0
    */
   @Override
-  public void shutdown() {}
+  public void shutdown() {
+    this.sparkContext = null; // release resource
+  }
 }