+++ adapter/spark/src/test/java/com/asml/apa/wta/spark/BaseSparkJobIntegrationTest.java
@@ -1,7 +1,10 @@
 package com.asml.apa.wta.spark;
 
+import com.asml.apa.wta.core.config.RuntimeConfig;
+import com.asml.apa.wta.core.model.enums.Domain;
 import com.asml.apa.wta.spark.datasource.SparkDataSource;
 import java.util.Arrays;
+import java.util.Map;
 import org.apache.spark.SparkConf;
 import org.apache.spark.api.java.JavaRDD;
 import org.apache.spark.api.java.JavaSparkContext;
@@ -11,13 +14,23 @@ import scala.Tuple2;
 
 public class BaseSparkJobIntegrationTest {
 
+  public static final String RESOURCE_PATH = "src/test/resources/wordcount.txt";
   protected SparkSession spark;
   protected SparkDataSource sut;
 
   protected JavaRDD<String> testFile;
 
+  RuntimeConfig fakeConfig;
+
   @BeforeEach
   void setup() {
+    fakeConfig = RuntimeConfig.builder()
+        .authors("Harry Potter")
+        .domain(Domain.SCIENTIFIC)
+        .description("Yer a wizard harry")
+        .events(Map.of("event1", "Desc of event1", "event2", "Desc of event2"))
+        .build();
+
     SparkConf conf = new SparkConf()
         .setAppName("SparkTestRunner")
         .setMaster("local[1]")
@@ -26,7 +39,7 @@ public class BaseSparkJobIntegrationTest {
     spark = SparkSession.builder().config(conf).getOrCreate();
     spark.sparkContext().setLogLevel("ERROR");
 
-    sut = new SparkDataSource(spark.sparkContext());
+    sut = new SparkDataSource(spark.sparkContext(), fakeConfig);
     String resourcePath = "src/test/resources/wordcount.txt";
     testFile = JavaSparkContext.fromSparkContext(spark.sparkContext()).textFile(resourcePath);
   }