+++ adapter/spark/src/test/java/com/asml/apa/wta/spark/listener/TaskLevelListenerTest.java
@@ -6,7 +6,6 @@ import static org.mockito.Mockito.when;
 
 import com.asml.apa.wta.core.model.Task;
 import java.util.Properties;
-import org.apache.spark.SparkContext;
 import org.apache.spark.executor.ExecutorMetrics;
 import org.apache.spark.executor.TaskMetrics;
 import org.apache.spark.scheduler.*;
@@ -16,9 +15,8 @@ import org.junit.jupiter.api.Test;
 import scala.collection.immutable.Seq;
 import scala.collection.mutable.ListBuffer;
 
-class TaskLevelListenerTest {
+class TaskLevelListenerTest extends BaseLevelListenerTest {
 
-  SparkContext mockedSparkContext;
   TaskLevelListener sut;
 
   TaskInfo testTaskInfo;
@@ -29,9 +27,7 @@ class TaskLevelListenerTest {
 
   @BeforeEach
   void setup() {
-    mockedSparkContext = mock(SparkContext.class);
-    when(mockedSparkContext.sparkUser()).thenReturn("testUser");
-    sut = new TaskLevelListener(mockedSparkContext);
+    sut = new TaskLevelListener(mockedSparkContext, fakeConfig);
 
     testTaskInfo = new TaskInfo(1, 0, 1, 50L, "testExecutor", "local", TaskLocality.NODE_LOCAL(), false);
 
@@ -61,8 +57,8 @@ class TaskLevelListenerTest {
 
     sut.onJobStart(new SparkListenerJobStart(1, 2L, stageBuffer.toList(), new Properties()));
     sut.onTaskEnd(taskEndEvent);
-    assertEquals(1, sut.getProcessedTasks().size());
-    Task curTask = sut.getProcessedTasks().get(0);
+    assertEquals(1, sut.getProcessedObjects().size());
+    Task curTask = sut.getProcessedObjects().get(0);
     assertEquals(1, curTask.getId());
     assertEquals("testtaskType", curTask.getType());
     assertEquals(50L, curTask.getSubmitTime());