+++ core/src/main/java/com/asml/apa/wta/core/utils/WtaUtils.java
@@ -1,6 +1,7 @@
 package com.asml.apa.wta.core.utils;
 
 import com.asml.apa.wta.core.config.RuntimeConfig;
+import com.asml.apa.wta.core.model.enums.Domain;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -8,6 +9,13 @@ import java.io.FileInputStream;
 import java.util.HashMap;
 import java.util.Map;
 
+/**
+ * Utility class for WTA.
+ *
+ * @author Henry Page
+ * @author Lohithsai Yadala Chanchu
+ * @since 1.0.0
+ */
 public class WtaUtils {
 
   private static final String CONFIG_DIR = "config.json";
@@ -37,19 +45,23 @@ public class WtaUtils {
       JsonNode resourceNode = rootNode.get("resourceSettings");
 
       String author = workloadNode.get("author").asText();
-      String domain = workloadNode.get("domain").asText();
+
+      Domain domain = Domain.extractAsEnum(workloadNode.get("domain").asText());
+
       String description = workloadNode.has("description")
           ? workloadNode.get("description").asText()
           : "";
       Map<String, String> events = resourceNode.has("events")
-          ? mapper.convertValue(resourceNode.get("events"), new TypeReference<Map<String, String>>() {})
+          ? mapper.convertValue(resourceNode.get("events"), new TypeReference<>() {})
           : new HashMap<>();
       configBuilder = configBuilder
-          .author(author)
+          .authors(author)
           .domain(domain)
           .description(description)
           .events(events);
-
+    } catch (EnumConstantNotPresentException e) {
+      throw new IllegalArgumentException(e.constantName()
+          + " is not a valid domain. It must be one of BIOMEDICAL, ENGINEERING, INDUSTRIAL, SCIENTIFIC.");
     } catch (Exception e) {
       throw new IllegalArgumentException(
           "The config file has missing/invalid fields or no config file was found");