+++ core/src/test/java/com/asml/apa/wta/core/ConfigReaderTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.asml.apa.wta.core.config.RuntimeConfig;
+import com.asml.apa.wta.core.model.enums.Domain;
 import com.asml.apa.wta.core.utils.WtaUtils;
 import java.util.HashMap;
 import org.junit.jupiter.api.Test;
@@ -12,8 +13,8 @@ public class ConfigReaderTest {
   @Test
   void readsConfigFileCorrectly() {
     RuntimeConfig cr = WtaUtils.readConfig("src/test/resources/testConfig.json");
-    assertThat(cr.getAuthor()).isEqualTo("Test Name");
-    assertThat(cr.getDomain()).isEqualTo("Test Domain");
+    assertThat(cr.getAuthors()).isEqualTo("Test Name");
+    assertThat(cr.getDomain()).isEqualTo(Domain.SCIENTIFIC);
     assertThat(cr.getDescription()).isEqualTo("Test Description");
     assertThat(cr.getEvents()).isEqualTo(new HashMap<String, String>() {
       {
@@ -26,8 +27,8 @@ public class ConfigReaderTest {
   @Test
   void readsConfigFileWhereTheDescriptionIsNotThere() {
     RuntimeConfig cr = WtaUtils.readConfig("src/test/resources/testConfigNoDesc.json");
-    assertThat(cr.getAuthor()).isEqualTo("Test Name");
-    assertThat(cr.getDomain()).isEqualTo("Test Domain");
+    assertThat(cr.getAuthors()).isEqualTo("Test Name");
+    assertThat(cr.getDomain()).isEqualTo(Domain.ENGINEERING);
     assertThat(cr.getDescription()).isEqualTo("");
     assertThat(cr.getEvents()).isEqualTo(new HashMap<String, String>() {
       {
@@ -40,8 +41,8 @@ public class ConfigReaderTest {
   @Test
   void readsConfigFileWhereTheEventsAreNotThere() {
     RuntimeConfig cr = WtaUtils.readConfig("src/test/resources/testConfigNoEvents.json");
-    assertThat(cr.getAuthor()).isEqualTo("Test Name");
-    assertThat(cr.getDomain()).isEqualTo("Test Domain");
+    assertThat(cr.getAuthors()).isEqualTo("Test Name");
+    assertThat(cr.getDomain()).isEqualTo(Domain.INDUSTRIAL);
     assertThat(cr.getDescription()).isEqualTo("Test Description");
     assertThat(cr.getEvents()).isEqualTo(new HashMap<String, String>());
   }